Ai/2D+자율주행

+++ src/main/java/com/Study/vibeclean/dto/mqtt/TelemetryMessage.java
@@ -16,14 +16,14 @@ public class TelemetryMessage {
 
     @Data
     public static class Position {
-        private int x;
-        private int y;
+        private float x;
+        private float y;
 
-        public int getX() {
+        public float getX() {
             return x;
         }
 
-        public int getY() {
+        public float getY() {
             return y;
         }
     }
